UPBGE: Implement filter cube map texture.
Previously only 2D texture were bindable before rendering a filter
thanks to the function KX_2DFilter.setTexture. But in some case the
user want to use a cube map (e.g world probe) in his shader.

To fill this gap a new python function is added: KX_2DFilter.setCubeMap
with the same argument than setTexture.

This function internally add a pair of values to the texture unordered
map stored in RAS_2DFilter, the first value is the binding category
and the second value the bind id.
Then in RAS_2DFilter we read this map to bind all teh textures at the
correct index (map key) and binding category.

Fix issue #529.