UPBGE: Fix crash on scene exit.
Previously the DAB (display arrray bucket) were owned by the mesh object
through the mesh material. The DAB need to be unregistered from the material
bucket before its deletion. But this was impossible because the scene which
own the bucket manager and so the material bucket was freed before freing the
DAB trough the meshes.

To fix this issue the order is inverted, the scene is freed after the meshes,
this cause no issue has no reading of the meshes and material are made in
the scene and object destruction. About deformers there's no problems too as
the bucket manager is freed after deleting all the object and so the deformers
in the scene.